non-coded/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
no-cache-dir==0.0.0
numpy==1.23.4
pip-autoremove==0.10.0
pip==22.3.1
protobuf==3.20.1
requests==2.28.2
setuptools==65.5.0
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
